requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
python-decouple
sqlalchemy
python-multipart